+++ packages/image_picker/example/lib/main.dart
@@ -34,6 +34,12 @@ class MyHomePage extends StatefulWidget {
 class _MyHomePageState extends State<MyHomePage> {
   Future<File> _imageFile;
 
+  void _onImageButtonPressed(ImageSource source) {
+    setState(() {
+      _imageFile = ImagePicker.pickImage(source: source);
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
     return new Scaffold(
@@ -41,26 +47,37 @@ class _MyHomePageState extends State<MyHomePage> {
         title: const Text('Image Picker Example'),
       ),
       body: new Center(
-          child: new FutureBuilder<File>(
-              future: _imageFile,
-              builder: (BuildContext context, AsyncSnapshot<File> snapshot) {
-                if (snapshot.connectionState == ConnectionState.done &&
-                    snapshot.error == null) {
-                  return new Image.file(snapshot.data);
-                } else if (snapshot.error != null) {
-                  return const Text('error picking image.');
-                } else {
-                  return const Text('You have not yet picked an image.');
-                }
-              })),
-      floatingActionButton: new FloatingActionButton(
-        onPressed: () {
-          setState(() {
-            _imageFile = ImagePicker.pickImage();
-          });
-        },
-        tooltip: 'Pick Image',
-        child: new Icon(Icons.add_a_photo),
+        child: new FutureBuilder<File>(
+          future: _imageFile,
+          builder: (BuildContext context, AsyncSnapshot<File> snapshot) {
+            if (snapshot.connectionState == ConnectionState.done &&
+                snapshot.error == null) {
+              return new Image.file(snapshot.data);
+            } else if (snapshot.error != null) {
+              return const Text('error picking image.');
+            } else {
+              return const Text('You have not yet picked an image.');
+            }
+          },
+        ),
+      ),
+      floatingActionButton: new Column(
+        mainAxisAlignment: MainAxisAlignment.end,
+        children: <Widget>[
+          new FloatingActionButton(
+            onPressed: () => _onImageButtonPressed(ImageSource.gallery),
+            tooltip: 'Pick Image from gallery',
+            child: new Icon(Icons.photo_library),
+          ),
+          new Padding(
+            padding: const EdgeInsets.only(top: 16.0),
+            child: new FloatingActionButton(
+              onPressed: () => _onImageButtonPressed(ImageSource.camera),
+              tooltip: 'Take a Photo',
+              child: new Icon(Icons.camera_alt),
+            ),
+          ),
+        ],
       ),
     );
   }