+++ packages/image_picker/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.4.0
+
+* **Breaking change**. The `source` parameter for the `pickImage` is now required. Also, the `ImageSource.any` option doesn't exist anymore.
+* Use the native Android image gallery for picking images instead of a custom UI. 
+
 ## 0.3.1
 
 * Bugfix: Android version correctly asks for runtime camera permission when using `ImageSource.camera`.